Add support for profile fields
This is client support for MSC4311, which allows us to (finally) set and
read custom profile fields. Currently, we only support the stable m.tz
timezone field.

These settings are hidden or disabled depending of it's supported or
allowed on your server. I also made sure to only make it writeable if
your server declares it so.

Fixes #661