Fix compact mode not filling all available space
This seems to be an unintentional change in compact mode caused by 054f87cae2cc09f92a2cbbf981366278accee95f,
where an if check for "straight line" or "fill width" mode ended up
being removed. But this was needed for availableWidth() to return the
correct width for compact mode, otherwise it got weirdly centered and
ended up with a limited width.